requirements.txt
requests==2.31.0
python-dotenv==1.0.0
gTTS==2.3.2
urllib3==2.0.7
flask==3.0.0
psycopg2-binary